Description
This app, when run should remind the client if there is a birthday on the list of birthday entered in the file. Perhaps I could add a gui that allows the user to enter the birthdays of their loved ones and then saved them in a list that it then reads. I also need to find a way to make it run once every 24 hours.