Create Book_Release_Radar
I've decided to make a program that keeps track of books I'm waiting for and notifies me on and a set amount of days prior to their release date. Once I've gotten that down I'll see if I can add features like adding new books to the databse manually or webscraping to see if my favorite authors have announced any new books recently and if so send the user a notification.